@@ -4,7 +4,7 @@ version: 1.0.2+3
 homepage: https://github.com/flutter/plugins/tree/master/packages/integration_test
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.13+hotfix.5"
 
 dependencies:
@@ -14,14 +14,14 @@ dependencies:
     sdk: flutter
   flutter_test:
     sdk: flutter
-  path: ^1.6.4
+  path: ^1.8.0
   # TODO(dnfield): This is a problem - flutter_driver and flutter_tools depend
-  # on this packkage, and so does integration_test. When this gets rev'd in the
+  # on this package, and so does integration_test. When this gets rev'd in the
   # SDK, it has to be rev'd here too so integration tests in the SDK can use it.
   vm_service: ">= 4.2.0 <7.0.0"
 
 dev_dependencies:
-  pedantic: ^1.8.0
+  pedantic: ^1.11.0
   mockito: ^4.1.1
 
 flutter:
