@@ -6,6 +6,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
+// ignore: import_of_legacy_library_into_null_safe
 import 'package:flutter_driver/flutter_driver.dart';
 
 import 'package:integration_test/common.dart';
@@ -21,16 +22,16 @@ String testOutputsDirectory =
 
 /// The callback type to handle [integration_test.Response.data] after the test
 /// succeeds.
-typedef ResponseDataCallback = FutureOr<void> Function(Map<String, dynamic>);
+typedef ResponseDataCallback = FutureOr<void> Function(Map<String, dynamic>?);
 
 /// Writes a json-serializable json data to to
 /// [testOutputsDirectory]/`testOutputFilename.json`.
 ///
 /// This is the default `responseDataCallback` in [integrationDriver].
 Future<void> writeResponseData(
-  Map<String, dynamic> data, {
+  Map<String, dynamic>? data, {
   String testOutputFilename = 'integration_response_data',
-  String destinationDirectory,
+  String? destinationDirectory,
 }) async {
   assert(testOutputFilename != null);
   destinationDirectory ??= testOutputsDirectory;
@@ -86,6 +87,6 @@ Future<void> integrationDriver({
 
 const JsonEncoder _prettyEncoder = JsonEncoder.withIndent('  ');
 
-String _encodeJson(Map<String, dynamic> jsonObject, bool pretty) {
+String _encodeJson(Map<String, dynamic>? jsonObject, bool pretty) {
   return pretty ? _prettyEncoder.convert(jsonObject) : json.encode(jsonObject);
 }
