@@ -17,7 +17,7 @@ const String _failureExcerpt = 'Expected: <false>\\n  Actual: <true>';
 void main() async {
   group('Integration binding result', () {
     test('when multiple tests pass', () async {
-      final Map<String, dynamic> results =
+      final Map<String, dynamic>? results =
           await _runTest('test/data/pass_test_script.dart');
 
       expect(
@@ -29,7 +29,7 @@ void main() async {
     });
 
     test('when multiple tests fail', () async {
-      final Map<String, dynamic> results =
+      final Map<String, dynamic>? results =
           await _runTest('test/data/fail_test_script.dart');
 
       expect(results, hasLength(2));
@@ -40,7 +40,7 @@ void main() async {
     });
 
     test('when one test passes, then another fails', () async {
-      final Map<String, dynamic> results =
+      final Map<String, dynamic>? results =
           await _runTest('test/data/pass_then_fail_test_script.dart');
 
       expect(results, hasLength(2));
@@ -53,7 +53,7 @@ void main() async {
 /// Runs a test script and returns the [IntegrationTestWidgetsFlutterBinding.result].
 ///
 /// [scriptPath] is relative to the package root.
-Future<Map<String, dynamic>> _runTest(String scriptPath) async {
+Future<Map<String, dynamic>?> _runTest(String scriptPath) async {
   final Process process =
       await Process.start(_flutterBin, ['test', '--machine', scriptPath]);
 
@@ -72,16 +72,16 @@ Future<Map<String, dynamic>> _runTest(String scriptPath) async {
               // Only interested in test events which are JSON.
             }
           })
-          .expand<Map<String, dynamic>>((dynamic json) {
+          .expand<Map<String, dynamic>?>((dynamic json) {
             return json is List<dynamic>
                 ? json.cast()
-                : <Map<String, dynamic>>[json as Map<String, dynamic>];
+                : <Map<String, dynamic>?>[json as Map<String, dynamic>];
           })
           .where((dynamic testEvent) =>
               testEvent != null && testEvent['type'] == 'print')
-          .map((dynamic printEvent) => printEvent['message'] as String)
-          .firstWhere((String message) =>
-              message.startsWith(_integrationResultsPrefix)))
+          .map((dynamic printEvent) => printEvent['message'] as String?)
+          .firstWhere((String? message) =>
+              message!.startsWith(_integrationResultsPrefix)))!
       .replaceAll(_integrationResultsPrefix, '');
 
   return jsonDecode(testResults);
