@@ -8,8 +8,8 @@ import 'package:integration_test/integration_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() async {
-  final IntegrationTestWidgetsFlutterBinding binding =
-      IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+  final IntegrationTestWidgetsFlutterBinding? binding =
+      IntegrationTestWidgetsFlutterBinding.ensureInitialized() as IntegrationTestWidgetsFlutterBinding?;
 
   testWidgets('passing test 1', (WidgetTester tester) async {
     expect(true, true);
@@ -21,6 +21,6 @@ void main() async {
 
   tearDownAll(() {
     print(
-        'IntegrationTestWidgetsFlutterBinding test results: ${jsonEncode(binding.results)}');
+        'IntegrationTestWidgetsFlutterBinding test results: ${jsonEncode(binding!.results)}');
   });
 }
